requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
matplotlib>=3.4.0
tqdm>=4.60.0
pillow>=8.0.0
scipy>=1.6.0
scikit-learn>=0.24.0
tensorboard>=2.5.0
einops>=0.4.0
accelerate>=0.12.0
datasets>=2.0.0